backend/langchain/pyproject.toml
[project]
name = "forecast"
version = "0.1.0"
description = "Urban Intelligence Platform"
authors = [
    { name = "Marc", email = "marcdasilva234@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.5",
    "celery>=5.5.3",
    "fastapi>=0.117.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.7",
    "langsmith>=0.4.31",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pgvector>=0.4.1",
    "playwright==1.58.0",
    "prophet>=1.1.7",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.36.0",
]

[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.2",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"../tools/web_scraping/Webscrape/*.py" = ["E402", "E701", "E741"]
